BigNumber and Ads
Updated BigNumber from other branch, bringing it into main. Also added samples of ad display to start implementing ads into the project

+++ Assets/Scripts/NPC/NPCStats.cs
@@ -185,7 +185,7 @@ public class NPCStats : MonoBehaviour
     //return the percent health left of the NPC as a float
     public float PercentHealth()
     {
-        return currentHealth / health;
+        return (float)(currentHealth / health);
     }
 
     //returns true if current health is at maximum
